requirements.txt
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.6.2
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==2.0.0
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
typing_extensions==4.12.2
tzdata==2024.1